Updates files so that files from the BlurParser repository are used.
All the files removed can now be found in the BlurParser repository.
In order to make the program work, just make a link or put the BlurParser
directory inside this repo's directory. The makefile will do the rest.

The SyntaxNode class was not removed as the Unify Algorithm app introduces
changes to this class which must be kept for the program to compile correctly.